+++ src/behavioral_patterns/observer/Subject.java
@@ -1,0 +1,7 @@
+package behavioral_patterns.observer;
+
+public interface Subject {
+    void addObserver(Observer observer);
+    void removeObserver(Observer observer);
+    void notifyObservers(String message);
+}