+++ src/behavioral_patterns/strategy/DifficultyStrategy.java
@@ -1,0 +1,13 @@
+package behavioral_patterns.strategy;
+
+import app.Assignment;
+
+import java.util.List;
+
+public class DifficultyStrategy implements PrioritizationStrategy {
+    @Override
+    public List<Assignment> prioritize(List<Assignment> assignments) {
+        System.out.println("Prioritizing by difficulty");
+        return assignments;
+    }
+}